Changed segment enums: easy to go to assembly address
Original segment enum assumed all indices were valid, when Pointer and
Temp only allowed 2 and 8 respectivly. Yeah, I could have thrown an
error when trying to do something with a wrong index, but it's cleaner
to have the enum never allow them and have the Reader throw the error if
a VM line tries to do it. All downstream code can now be happy.