Assembler now allows for multiple labels pointing to same command
Initial design did not allow this, as I couldn't think of a reason why
two labels would need to follow one another. Them while coding the VM, I
had a label at which code execution starts, followed by another label
for the first function. So now I can think of a reason to do and I can't
think of a reason not to do it.