The stack pointer works in the nested functions test!
What about all the other memory locations. Ok, I don't know yet. There
was a bug in the VM `return` implementation. It only happened when ARG
was 0 (as in `call foo 0`) in the function call.I tested on ARG == 3.
At 0, the return  address on the frame was being overwritten on the
`return` command.